fix pr comment about mainEvents by moving MainEvents to his own class file, and save MainActivity in is own package inside UI
