+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,8 @@
         android:name=".NewKeepItDroidApp"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.activities.main.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.NewKeepItDroid">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
