+++ app/src/main/java/es/infolojo/newkeepitdroid/ui/activities/main/events/MainEvents.kt
@@ -1,0 +1,7 @@
+package es.infolojo.newkeepitdroid.ui.activities.main.events
+
+import es.infolojo.newkeepitdroid.ui.screens.vo.UIMessagesVO
+
+sealed interface MainEvents {
+    data class ShowMessage(val message: UIMessagesVO) : MainEvents
+}
